feat: adicionado configuração automática do numero de concurso e para…

+++ src/main/java/com/estudo/megasema/dto/resultados/SaveResultadoDto.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class SaveResultadoDto {
 
     private String nome;
-    private int numeroConcurso;
+    private Long numeroConcurso;
     private String dataConcurso;
     private Boolean acumulou;
     private BigDecimal valorAcumulado;
